refactor(ChessGame): dao -> domain.board로 패키지 이동
