+++ src/test/java/chess/dao/MovementDAOTest.java
@@ -1,5 +1,6 @@
 package chess.dao;
 
+import chess.domain.board.ChessGame;
 import chess.domain.piece.property.Team;
 import chess.domain.position.Position;
 import org.junit.jupiter.api.Assertions;
